requirements.txt
streamlit
google-generativeai
python-env